A batch of changes that have piled up:
- Cleaner build system organization (1 SConscript file per directory instead
  of a single huge file)
- Robustness improvements to the KD-tree construction code (now does a better
  job at handling degenerate triangles)
- Switched to an epsilon-free KD-tree traversal loop using Havran's T_{AB}^rec
  algorithm
- Generalization of the KD-tree construction code (now supports plugging in
  arbitrary tree construction heuristics)
- Addition of some utility code (LRU cache, adaptive Gauss-Lobatto
  quadrature, etc.)
- Generic dense matrix class that supports arbitrary dimensions
- Pixel traversal within image blocks now uses a space-filling curve
  ordering
- Added support for several noise functions from PBRT
- Cleanups of various top-level interfaces (Luminaires, Phase functions,
  BSDFs, Participating media, etc.)
- Robustness improvements to the participating media code (faster + now
  does a better job at dielectric boundaries)
- Added a basic tonemapping utility (can be invoked via mtsutil)

+++ include/mitsuba/core/shvector4d.h
@@ -74,14 +74,14 @@ public:
 
 		for (int i=0; i<=res; ++i) {
 			Float theta = hExt*i, cosTheta = std::cos(theta);
-			Float weightExt = (i & 1) ? 4 : 2;
+			Float weightExt = (i & 1) ? (Float) 4 : (Float) 2;
 			if (i == 0 || i == res)
 				weightExt = 1;
 			SLog(EInfo, "4D projection: %i%% done", (int) (100*i/(Float) res));
 
 			for (int j=0; j<=res*2; ++j) {
 				Float phi = hInt*j;
-				Float weightInt = (j & 1) ? 4 : 2;
+				Float weightInt = (j & 1) ? (Float) 4 : (Float) 2;
 				if (j == 0 || j == 2*res)
 					weightInt = 1;
